update test.js for covering more cases
Updating test.js covering. The current set of tests is not full.
Just because tests f(1, 3) and f(1, 5) start from 1 (the neutral element for multiplying), they can't prevent newbies to start their loops from the second element of the slice. 
E.g. f(1, 5) = f(2, 5). But!(2,5) != (3,5) -- and cases as this was not covered.
Here is an example of a wrong function that passes through all tests.

const multiplyNumbersFromRange = (a, b) => {
  let i = a + 1;  // here we can start from second element and it will be okay
  let mult = 1;
  if (b-a != 0) {
    while (i <= b) {
      mult = mult * i;
      i = i + 1;
    } 
  }else {
      mult = a;
	}
  return mult;
}